Asteroid Fix
whitelisting ore for the asteroids in the advrocketry.cfg config file actually doesn't spawn the new ore due to all of them having no "chance" to generate. Added in the new ore to asteroids and added medium, large, and monster sized asteroids while eliminating the iridium rich asteroid. For whatever reason the default only had two sizes which makes it lackluster for the amount of work needed to mine an asteroid. 

Chances of both ore and asteroid generation may need to be tuned.